Refactor test to appear only in user sees job spec
